Fixed id reference bug. fixed percentage bug.
there was a bug where ids would end up being identical in two spots
causing two or more rows to collapse/expand when one was clicked. I
fixed this using randomly generated ids.

The graph at the top of the page often times would not display the
correct percentage (say in the case of 999 pass 1 fail the percent
showed 100%). I fixed this by using the Math.floor method to ensure it
would not round up.